Move LiveLeaderboard component into Sidebar
