+++ src/components/Exchange/SwapBox.js
@@ -79,7 +79,7 @@ import longImg from "../../img/long.svg";
 import shortImg from "../../img/short.svg";
 import swapImg from "../../img/swap.svg";
 import { useUserReferralCode } from "../../Api/referrals";
-import {  getMaxLeverage , LeverageInput } from "./LeverageInput";
+import { getMaxLeverage, LeverageInput } from "./LeverageInput";
 import { REFERRAL_CODE_KEY } from "../../config/localstorage";
 
 const SWAP_ICONS = {
@@ -149,7 +149,6 @@ export default function SwapBox(props) {
     isPluginApproving,
     isPositionRouterApproving,
     trackAction,
-    setIsLeaderboardVisible,
     updateLeaderboardOptimistically,
   } = props;
 
@@ -341,7 +340,7 @@ export default function SwapBox(props) {
             ></Tooltip>
           </div>
         </div>
-      )
+      );
     } else if (isLong) {
       return (
         <div className="Exchange-info-row">
@@ -379,11 +378,13 @@ export default function SwapBox(props) {
                 return (
                   <>
                     {`Max ${fromTokenInfo.symbol} in: `}
-                    {formatAmount(maxFromTokenIn, fromTokenInfo.decimals, 0, true)} {fromTokenInfo.symbol} (${formatAmount(maxFromTokenInUSD, USD_DECIMALS, 0, true)})
+                    {formatAmount(maxFromTokenIn, fromTokenInfo.decimals, 0, true)} {fromTokenInfo.symbol} ($
+                    {formatAmount(maxFromTokenInUSD, USD_DECIMALS, 0, true)})
                     <br />
                     <br />
                     {`Max ${toTokenInfo.symbol} out: `}
-                    {formatAmount(maxToTokenOut, toTokenInfo.decimals, 0, true)} {toTokenInfo.symbol} (${formatAmount(maxToTokenOutUSD, USD_DECIMALS, 0, true)})
+                    {formatAmount(maxToTokenOut, toTokenInfo.decimals, 0, true)} {toTokenInfo.symbol} ($
+                    {formatAmount(maxToTokenOutUSD, USD_DECIMALS, 0, true)})
                     <br />
                   </>
                 );
@@ -391,11 +392,10 @@ export default function SwapBox(props) {
             />
           </div>
         </div>
-      )
+      );
     } // else
-    return null
+    return null;
   };
-
 
   const fromBalance = fromTokenInfo ? fromTokenInfo.balance : bigNumberify(0);
   const toBalance = toTokenInfo ? toTokenInfo.balance : bigNumberify(0);
@@ -852,7 +852,7 @@ export default function SwapBox(props) {
       toTokenInfo.poolAmount &&
       toTokenInfo.bufferAmount.gt(toTokenInfo.poolAmount.sub(toAmount))
     ) {
-      console.log(toTokenInfo.poolAmount.toString(), toTokenInfo.bufferAmount.toString())
+      console.log(toTokenInfo.poolAmount.toString(), toTokenInfo.bufferAmount.toString());
       return ["Insufficient liquidity: exceeds buffer"];
     }
 
@@ -1453,7 +1453,6 @@ export default function SwapBox(props) {
       .finally(() => {
         setIsSubmitting(false);
         setIsPendingConfirmation(false);
-        setIsLeaderboardVisible(true);
       });
   };
 
@@ -1505,7 +1504,6 @@ export default function SwapBox(props) {
       .finally(() => {
         setIsSubmitting(false);
         setIsPendingConfirmation(false);
-        setIsLeaderboardVisible(true);
       });
   };
 
@@ -1642,8 +1640,6 @@ export default function SwapBox(props) {
       .finally(() => {
         setIsSubmitting(false);
         setIsPendingConfirmation(false);
-        setIsLeaderboardVisible(true);
-        setIsLeaderboardVisible(true);
         updateLeaderboardOptimistically(fromUsdMin);
       });
   };
@@ -2241,7 +2237,13 @@ export default function SwapBox(props) {
         )}
         {(isLong || isShort) && (
           <div className="Exchange-leverage-box">
-            <LeverageInput value={leverageOption} onChange={setLeverageOption} max={getMaxLeverage(toToken.symbol)} min={1.1} step={0.01} />
+            <LeverageInput
+              value={leverageOption}
+              onChange={setLeverageOption}
+              max={getMaxLeverage(toToken.symbol)}
+              min={1.1}
+              step={0.01}
+            />
             {isShort && (
               <div className="Exchange-info-row">
                 <div className="Exchange-info-label">Profits In</div>
