+++ src/views/Exchange/Exchange.js
@@ -30,14 +30,12 @@ import {
   useAccountOrders,
   getPageTitle,
   useLocalStorageSerializeKey,
-  getTracerServerUrl,
 } from "../../Helpers";
 import { getConstant } from "../../Constants";
-import { approvePlugin, useMarketMakingFeesSince, useFeesSince, useChartPrices } from "../../Api";
+import { approvePlugin, useChartPrices } from "../../Api";
 
 import { getContract } from "../../Addresses";
 import { getTokens, getToken, getWhitelistedTokens, getTokenBySymbol } from "../../data/Tokens";
-import { getFeeHistory, SECONDS_PER_WEEK } from "../../data/Fees";
 
 import Reader from "../../abis/ReaderV2.json";
 import VaultV2 from "../../abis/VaultV2.json";
@@ -57,12 +55,8 @@ import Tab from "../../components/Tab/Tab";
 import "./Exchange.css";
 import SEO from "../../components/Common/SEO";
 import { ExchangeHeader } from "../../components/Exchange/ExchangeHeader";
-import FeeUpdateModal from "../../components/Modal/FeeUpdateModal";
-import LiveLeaderboard from "./LiveLeaderboard";
-import { getUnclaimedFees } from "../Dashboard/DashboardV2";
 
 const { AddressZero } = ethers.constants;
-
 const PENDING_POSITION_VALID_DURATION = 600 * 1000;
 const UPDATED_POSITION_VALID_DURATION = 60 * 1000;
 
@@ -368,12 +362,11 @@ export const Exchange = forwardRef((props, ref) => {
     trackAction,
     sidebarVisible,
     analytics,
+    updateLeaderboardOptimistically,
   } = props;
 
   const [pendingPositions, setPendingPositions] = useState({});
   const [updatedPositions, setUpdatedPositions] = useState({});
-
-  const [isLeaderboardVisible, setIsLeaderboardVisible] = useState(false);
 
   useEffect(() => {
     window.scrollTo(0, 0);
@@ -908,69 +901,6 @@ export const Exchange = forwardRef((props, ref) => {
     setChartToken(tmp);
   }, [swapOption, fromToken, toToken, chainId]);
 
-  const { data: currentRewardRound, error: failedFetchingRoundRewards } = useSWR(
-    [getTracerServerUrl(chainId, "/tradingRewards"), "latest"],
-    {
-      fetcher: (url, round) => fetch(`${url}&round=${round}`).then((res) => res.json()),
-    }
-  );
-  const whitelistedTokenAddresses = whitelistedTokens.map((token) => token.address);
-  const stableTokens = whitelistedTokens.filter((t) => t.isStable);
-
-  const { data: fees } = useSWR([`Dashboard:fees:${active}`, chainId, readerAddress, "getFees", vaultAddress], {
-    fetcher: fetcher(library, Reader, [whitelistedTokenAddresses]),
-  });
-
-  const feeHistory = getFeeHistory(chainId);
-
-  const from = feeHistory[0]?.to;
-  const to = from + SECONDS_PER_WEEK * 2;
-  const currentMMFees = useMarketMakingFeesSince(chainId, from, to, stableTokens);
-  const currentGraphFees = useFeesSince(chainId, from, to);
-  const currentUnclaimedFees = getUnclaimedFees(whitelistedTokenAddresses, infoTokens, fees);
-  let totalCurrentFees, currentFees, fivePercentOfFees;
-  if (currentUnclaimedFees && currentGraphFees) {
-    currentFees = currentUnclaimedFees.gt(currentGraphFees) ? currentUnclaimedFees : currentGraphFees;
-  }
-
-  if (currentFees && currentMMFees) {
-    totalCurrentFees = currentFees.add(currentMMFees);
-    fivePercentOfFees = totalCurrentFees.mul(5).div(100);
-  }
-
-  const [leaderboardData, setLeaderboardData] = useState([]);
-
-  // Get volume, position and reward from user round data
-  const userPosition = useMemo(() => {
-    if (!leaderboardData) {
-      return undefined;
-    }
-    const leaderBoardIndex = leaderboardData?.findIndex(
-      (trader) => trader.user_address.toLowerCase() === account?.toLowerCase()
-    );
-    return leaderBoardIndex + 1;
-  }, [account, leaderboardData]);
-
-  useEffect(() => {
-    if (!currentRewardRound || !!currentRewardRound?.message) {
-      return undefined;
-    }
-    const traders = currentRewardRound.rewards
-      ?.sort((a, b) => b.volume - a.volume)
-      .map((trader) => {
-        const positionReward = ethers.BigNumber.from(trader.reward);
-        const degenReward = ethers.BigNumber.from(trader.degen_reward);
-
-        return {
-          ...trader,
-          totalReward: positionReward.add(degenReward),
-          positionReward,
-          degenReward,
-        };
-      }); // Sort traders by highest to lowest in volume
-    setLeaderboardData(traders);
-  }, [currentRewardRound]);
-
   if (!getToken(chainId, toTokenAddress)) {
     return null;
   }
@@ -979,33 +909,8 @@ export const Exchange = forwardRef((props, ref) => {
     setFromTokenAddress(swapOption, token.address);
   };
 
-  const updateLeaderboardOptimistically = (volumeToAdd) => {
-    const _leaderboardData = [...leaderboardData];
-    const currentUser = _leaderboardData.find(
-      (trader) => trader?.user_address?.toLowerCase() === account?.toLowerCase()
-    );
-    if (currentUser) {
-      currentUser.volume = ethers.BigNumber.from(currentUser.volume).add.ethers.BigNumber.from(volumeToAdd).toString();
-    } else {
-      _leaderboardData.push({
-        user_address: account,
-        volume: volumeToAdd,
-        reward: 0,
-        degen_reward: 0,
-      });
-    }
-    setLeaderboardData(_leaderboardData);
-  };
-
   return (
     <>
-      <LiveLeaderboard
-        userPosition={userPosition}
-        leaderboardData={leaderboardData}
-        fivePercentOfFees={fivePercentOfFees}
-        isVisible={isLeaderboardVisible}
-        setIsVisible={setIsLeaderboardVisible}
-      />
       <SEO description="Trade with liquidity, leverage, low fees. Trade with Mycelium. Trade Perpetual Swaps and Perpetual Pools on Ethereum scaling solution, Arbitrum with liquid markets for BTC, ETH, LINK, UNI, CRV, FXS, & BAL." />
       <div className="Exchange default-container">
         <div className="Exchange-content">
@@ -1111,7 +1016,6 @@ export const Exchange = forwardRef((props, ref) => {
               totalTokenWeights={totalTokenWeights}
               usdgSupply={usdgSupply}
               trackAction={trackAction}
-              setIsLeaderboardVisible={setIsLeaderboardVisible}
               updateLeaderboardOptimistically={updateLeaderboardOptimistically}
             />
             <div className="Exchange-wallet-tokens">
